adminrag/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
chromadb
sentence-transformers
openpyxl
httpx
pypdf
pytesseract
pillow
pdf2image
